// client/src/types/types.ts
export interface WeatherData {
  coord: {
    lon: number;
    lat: number;
  };
  weather: {
    id: number;
    main: string;
    description: string;
    icon: string;
  }[];
  base?: string;
  main: {
    temp: number;
    feels_like: number;
    temp_min: number;
    temp_max: number;
    pressure: number;
    humidity: number;
    sea_level?: number;
    grnd_level?: number;
  };
  visibility: number;
  wind: {
    speed: number;
    deg: number;
    gust?: number;
  };
  clouds: {
    all: number;
  };
  dt: number;
  sys: {
    type?: number;
    id?: number;
    country: string;
    sunrise: number;
    sunset: number;
  };
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

export interface ForecastWeatherData {
  cod: number;
  message: number;
  cnt: number;
  list: {
    dt: number;
    main: {
      temp: number;
      feels_like: number;
      temp_min: number;
      temp_max: number;
      pressure: number;
      sea_level: number;
      grnd_level: number;
      humidity: number;
      temp_kf: number;
    };
    weather: {
      id: number;
      main: string;
      description: string;
      icon: string;
    }[];
    clouds: {
      all: number;
    };
    wind: {
      speed: number;
      deg: number;
      gust: number;
    };
    visibility?: number;
    pop?: number;
    sys: {
      pod: string;
    };
    dt_txt: string;
  }[];
}

export interface SearchBarProps {
  setCurrentWeather: (data: WeatherData) => void;
  setForecastData: (data: ForecastWeatherDataTrim[]) => void;
}

export interface WeatherCardProps {
  weatherCardData: WeatherData;
}

export interface ForecastProps {
  forecastWeatherData: ForecastWeatherDataTrim[];
}

export interface ForecastWeatherDataTrim {
  date: string;
  icon: string;
  temperature: number;
  weather: string;
}

export interface FavouritesProps {
  favoriteCities: string[];
  setFavorites: (favorites: string[] | ((prevFavorites: string[]) => string[])) => void;
}